refactor(gateway): align inbound queue module layout
