+++ tests/setup/reset-inbound-queue.ts
@@ -1,5 +1,5 @@
 import { afterEach } from "vitest";
-import { resetInboundSessionQueueForTest } from "../../src/inbound-session-queue";
+import { resetInboundSessionQueueForTest } from "../../src/gateway/inbound-session-queue";
 
 // The inbound session-queue holds module-level state (the per-conversation
 // promise-chain map). Reset it after every test so a handler that hangs or
